refactor: 주문 상품 반정규화 및 상품 통계 테이블 생성

+++ src/main/java/com/prgrms/nabmart/domain/event/service/response/FindEventDetailResponse.java
@@ -19,7 +19,7 @@ public record FindEventDetailResponse(
     }
 
     public record EventItemResponse(Long itemId, String name, int price, int discount,
-                                    int reviewCount, int like, double rate) {
+                                    long reviewCount, long like, double rate) {
 
     }
 }
